fix: merge files with Rosponse Types
