+++ src/components/view/appView.ts
@@ -1,7 +1,6 @@
 import News from './news/news';
 import Sources from './sources/sources';
-import '../../types/RespNews';
-import '../../types/RespSources';
+import '../../types/ResponseTypes';
 
 export class AppView {
     readonly news: News;
