--- src/types/RespSources.ts
@@ -1,14 +1,0 @@
-type RespSources = {
-  status: string,
-  sources: Array<Source>
-}
-
-type Source = {
-  id: string,
-  name: string,
-  description: string,
-  url: string,
-  category: string,
-  language: string,
-  country: string,
-}
