+++ src/types/ResponseTypes.ts
@@ -1,5 +1,5 @@
 type RespNews = {
-  status: string,
+  status: Status,
   totalResults: number,
   articles: Array<Article>,
 };
@@ -18,4 +18,24 @@ type Article = {
 type ArticleSource = {
   id: string,
   name: string,
+}
+
+type RespSources = {
+  status: Status,
+  sources: Array<Source>
+}
+
+type Source = {
+  id: string,
+  name: string,
+  description: string,
+  url: string,
+  category: string,
+  language: string,
+  country: string,
+}
+
+enum Status {
+  ok,
+  error
 }
